Added user APIs

+++ src/main/java/com/codopedia/journalApp/service/JournalEntryService.java
@@ -1,6 +1,7 @@
 package com.codopedia.journalApp.service;
 
 import com.codopedia.journalApp.entity.JournalEntry;
+import com.codopedia.journalApp.entity.User;
 import com.codopedia.journalApp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,30 +16,63 @@ public class JournalEntryService {
     @Autowired
     private JournalEntryRepository journalEntryRepository ;
 
-    public void createJournal(JournalEntry journalEntry){
-        journalEntry.setDate(LocalDateTime.now());
-        journalEntryRepository.save(journalEntry);
+    @Autowired
+    private UserService userService;
+
+    public Boolean createJournalForUser(JournalEntry journalEntry, String userName){
+        try{
+            User user = userService.findByUserName(userName);
+            journalEntry.setDate(LocalDateTime.now());
+            JournalEntry createdJournal = journalEntryRepository.save(journalEntry);
+            user.getJournalEntryList().add(createdJournal);
+            userService.createUser(user);
+            return true;
+        }
+        catch (Exception ex){
+            return false;
+        }
     }
 
-    public List<JournalEntry> getAllJournal(){
-        return journalEntryRepository.findAll();
+
+    public List<JournalEntry> getAllJournalOfUser(String userName){
+        try{
+            User user = userService.findByUserName(userName);
+            return user.getJournalEntryList();
+        }
+        catch (Exception ex){
+            return null;
+        }
     }
 
     public JournalEntry getJournalByID(ObjectId id){
         return journalEntryRepository.findById(id).orElse(null);
     }
 
-    public void deleteByID(ObjectId id){
-        journalEntryRepository.deleteById(id);
+    public boolean deleteByID(String userName, ObjectId id){
+        User user = userService.findByUserName(userName);
+        JournalEntry journal = this.getJournalByID(id);
+        if(journal!=null &&  user!=null){
+            try{
+                user.getJournalEntryList().removeIf(x-> x.getId().equals(id));
+                userService.createUser(user);
+                journalEntryRepository.deleteById(id);
+                return true;
+            }
+            catch (Exception ex){
+                return false;
+            }
+        }
+        return false;
     }
 
-    public JournalEntry updateById(ObjectId id, JournalEntry journalEntry) {
+    public JournalEntry updateById(String userName, ObjectId id, JournalEntry journalEntry) {
         JournalEntry oldEntry = this.getJournalByID(id);
         if(oldEntry!=null){
             oldEntry.setTitle(journalEntry.getTitle()!=null && !journalEntry.getTitle().isEmpty() ? journalEntry.getTitle() : oldEntry.getTitle());
             oldEntry.setContent(journalEntry.getContent()!=null && !journalEntry.getContent().isEmpty() ? journalEntry.getContent() : oldEntry.getContent());
             journalEntryRepository.save(oldEntry);
+            return oldEntry;
         }
-        return oldEntry;
+        return null;
     }
 }
